refactor(client): adopt MVP pattern
